+++ src/read_hdf5_matrix.cpp
@@ -9,10 +9,18 @@
 #include "tatami/ext/HDF5DenseMatrix.hpp"
 #include "tatami/ext/HDF5CompressedSparseMatrix.hpp"
 
-NumericMatrix read_hdf5_matrix(std::string path, std::string name, bool layered) {
+struct Hdf5MatrixDetails {
     bool is_dense;
     bool csc = true;
     size_t nr, nc;
+};
+
+Hdf5MatrixDetails extract_hdf5_matrix_details_internal(const std::string& path, const std::string& name, bool do_dense) {
+    Hdf5MatrixDetails output;
+    auto& is_dense = output.is_dense;
+    auto& csc = output.csc;
+    auto& nr = output.nr;
+    auto& nc = output.nc;
 
     try {
         H5::H5File handle(path, H5F_ACC_RDONLY);
@@ -65,10 +73,41 @@ NumericMatrix read_hdf5_matrix(std::string path, std::string name, bool layered)
             } else {
                 throw std::runtime_error("expected a 'shape' attribute or dataset");
             }
+        } else if (do_dense) {
+            auto dhandle = handle.openDataSet(name);
+            auto dspace = dhandle.getSpace();
+            if (dspace.getSimpleExtentNdims() != 2) {
+                throw std::runtime_error("expected a 2-dimensional dataset for a dense matrix");
+            }
+
+            hsize_t dims[2];
+            dspace.getSimpleExtentDims(dims);
+            nr = dims[1]; // transposed deliberately, as HDF5 rows are typically samples => array columns.
+            nc = dims[0];
         }
     } catch (H5::Exception& e) {
         throw std::runtime_error(e.getCDetailMsg());
     }
+
+    return output;
+}
+
+void extract_hdf5_matrix_details(std::string path, std::string name, uintptr_t ptr) {
+    auto details = extract_hdf5_matrix_details_internal(path, name, true);
+    auto output = reinterpret_cast<int32_t*>(ptr);
+    output[0] = details.is_dense;
+    output[1] = details.csc;
+    output[2] = details.nr;
+    output[3] = details.nc;
+    return;
+}
+
+NumericMatrix read_hdf5_matrix(std::string path, std::string name, bool layered) {
+    auto details = extract_hdf5_matrix_details_internal(path, name, false);
+    const auto& is_dense = details.is_dense;
+    const auto& csc = details.csc;
+    const auto& nr = details.nr;
+    const auto& nc = details.nc;
 
     std::shared_ptr<tatami::Matrix<int, int> > mat;
     try {
@@ -91,6 +130,7 @@ NumericMatrix read_hdf5_matrix(std::string path, std::string name, bool layered)
  */
 EMSCRIPTEN_BINDINGS(read_hdf5_matrix) {
     emscripten::function("read_hdf5_matrix", &read_hdf5_matrix);
+    emscripten::function("extract_hdf5_matrix_details", &extract_hdf5_matrix_details);
 }
 /**
  * @endcond